Refactor delete_attendee ticket and move to ui folder. Create new concerts ui file. Refactor view_all_concerts
